Found and fixed timer bug. For some reason the AST generates spurious interrupts that do not relate to the ARn.VALUE register. It could be there are other ARn registers in use and I need to demultiplex? But regardless the issue was that sometimes interrupts would fire early and cause timers to fire early. This would then screw up the arithmetic to figure out what to do. The simple fix is to check that the interrupt has fired within a few ticks of the expected time. If not, schedule another interrupt. If so, handle the timer.
To test the timers, uncomment the lines in platform/storm/lib.rs
with 'periodic'. Connect the console. You should see output that
shows

tick #0x0002
   now: 0xC109
  last: 0xC10B
  when: 0xC90B
  next: 0xC307

where tick is which timer (there are two, with different periods),
now is the current time, last is the time this timer was scheduled
for, when is the time of the next firing of this timer, and next is
the value the interrupt will next fire at. For example, in the
above data you can see that there is an outstanding timer for 0xC307,
so 200 ticks in the future. The registered timer for 0xC90B is later
than this, so the compare at 0xC307 remains.