Add Shared/Private phantom type to AppPtr
Indicates whether an AppPtr (or AppSlice) was allocated and given
exclusively (i.e. there are no other references to that memory,
including by the app), or shared with the owner (e.g. passed from the
app). If `Shared`, the owner should not trust the integrity of the
memory -- can use it as a buffer to copy data to the app.