First attempt at using the Timer interface
Need to propogate 'static up to the syscall interface, plus the client
interface is a bit over-complex (need to two things, a Timer mux _and_
a Request, even though we never actually do anything with the request in
the client except pass it along). The request has to be in an Option.