Eliminate most static constraints for Timer/Alarm
Pushed all static constraints on self into VirtualTimer where it can't
be dealt with specially (possibly unsafely, but at least contained). One
thing this meant was change the Alarm trait to not take a `Request` in
`set_alarm`. This is OK, because the AST does do any sort of queueing
anyway, so the callback should be configured once at initialization.