VirtualTimer complete
Eliminate any reliance on `static for self in clients of the
VirtualTimer. Using an unsafe pointer to the TimerRequest, although it
seems like it should be possible without, since the reference would be a
'static one anyway. But seems OK for now.

We enforce that it was generated from a 'static reference by only
accepting those in `new`.