Reading from MISO works; something wrong with loopback setting in SPI mode register, so tested with a simple wiring. Behavior of the driver test (written in platform/storm/lib.rs) is it enters an infinite loop that writes 5 bytes to the SPI and pauses. The 5 bytes are 0x55, a counter, the byte read on the first write, the byte read on the second write, and 0. So if everything is working correctly it should be
  0x55 counter   0x55 counter   0
  0x55 counter+1 0x55 counter+1 0, etc.

To observe this, hook your logic analyzer up to pins 8 and 12 on the 22
pin header (next to "FIRESTORM 1.3"). Pin 8 is clock, pin 12 is MOSI.

Next step is testing the buffer-level read/write interface. This
interface uses a callback, but the first implementation will be
blocking, with a callback issued before the call returns (ew). After
that will be DMA.