Fix app memory alignment bug
Prior to this commit we were relying on LLVM to layout the apps'
memories aligned by simply initializing them as a mutable static slice
(in `src/process/process.rs`).

Unfortunately, while the app memories should be 8-byte aligned, LLVM
only enforces 4-byte alignment (at least for the cortex-m4), which makes
the memory layout sensitive to subtle changes in the other static
variables declared in the kernel (e.g. their size as well as which order
they are declared in).

Instead, we declare that slice of memory as belonging to the linker
section ".app_memory" and expect the chip's linker script to place that
memory in an appropriate place in zeroed out ram and align it
approriately (8-bytes on cortex-m4s at least).