Fix dependencies on syscalls.S
`syscalls.S` is architecture specific, which is why it's in the `arch`
subdirectory. However, it's only used by applications. Moreover,
appications don't use any of the other code in `arch.o` (like context
switching code). So it's not approriate for all of those to get compiled
into a single `arch.o` file.

This commit leaves source files as they are, but changes the build
system to include syscalls.S in applications, but not in the kernel, by
creating a new target: `build/apps/arch.o`.

Also reorganized the dependencies between Makefiles in the top-level
Makefile to make everything work (mostly to get the ARCH variable
defined by the time we get to application Makefiles)