Virtual timer had incorrect check for fired alarm
The virtual timer was checking if an alarm was _between_ the previous
value of the alarm and the current one to decide whether to fire it.
That is obviously the exact opposite of what we want... It was only
working because the checks were >= and the RC clock is slow enough that
we didn't _pass_ the alarm time by the time the handler was called.