nRF51822: Auto-generate peripheral interrupt vectors
The entries are generated automatically by nRF51_codegen.py based on SVD
models provided by the cmsis-svd project:
https://github.com/posborne/cmsis-svd

In theory, the same approach can be applied to other chips, as long as
SVD files are present on the cmsis-svd repository. If they are not
present but can be found in some SDK, they should be contributed to the
project via pull requests.