TMP006 driver and application
TMP006 temperature sensor driver with interface for reading
synchronously or asynchronously. Floating point math for sensor data is
performed in the kernel driver.

Application to demonstrate the TMP006 driver both synchronously and
asynchronously. Temperature readings are printed to the console in real
time.

Enable software-based floating point operations in the kernel. Wrappers
for libm functions to be used in the kernel are contained in
`common/math.rs`.

Various improvements for debugging driver/application development
 * keep .elf and .o files
 * create .lst disassembly files for user application and kernel
 * create .Map linker files for user application and kernel
 * `make listen` builds, loads, and connects to serial

Conflicts:
	src/apps/c_sync/Makefile.mk
	src/apps/libs/Makefile.mk
	src/common/lib.rs