tmp006: use more better floating point operations
Previously, the TMP006 driver used `pow` and `powi` functions. This
commit replace these with explicit multiplies and sqrt instead. These are
better choices for both code-size and efficiency.

On hardware with FPUs, both multiplication and sqrt are supported in the
hardware, while pow is still a library function. This saves about 7K of
code. Even without an FPU, using pow incurs about a 4K code size cost
relative to sqrt and explicit multiplies.