Pull apps out of main build system
The long-term goal is to separate the kernel and app build system
completely, with the caveat of the app build system can call the
kernel build as a first pre-req to make sure things are up to date
during development. This is most of the way there. It also treats
every application as a separate application, which I believe makes
more sense than a monolithic make system for all apps. The makefile
in the root of apps will recursively build each app.

Caveats:
  o The rust example app doesn't currently work, but it should really
    build with cargo instead of make anyway, so we'll revisit that later
  o The separation from the kernel tree is imperfect, but this is a
    good checkpoint where most things work correctly