SAM4L: Don't return buffer in DMA xfer_done
At least in the I2C case, xfer_done isn't where we want to do the logic
to get the buffer back. This decouples getting the signal from the DMA
from getting back control of the buffer. Instead, clients call
`abort_xfer` to stop the transfer (if it happens to still be going on)
and get the buffer back.