I2C: Progress on repeated starts
Order of setting next command and enabling the I2C seems to matter.
When the order is correct, we get a CCOMP as expected from the first
command and can at least set the DMA for the second with some transfer
happening. Not yet getting signal for the completion of the second
command.