I2C: Working repeated starts!!
The storm's i2c_dummy test module successfully uses the TMP006 as well
as the Light sensor with continuous readings using repeated starts. The
accelerometer is still not coopoerating and the behavior looks
suspiciously like
not-a-software-problem-even-though-it-probably-is-because-it-almost-always-is.

Specifically,the accelerometer works on a power cycle for a few commands
after activating it and then just bricks. Nothing else over I2C works
afterwords and the I2C driver is even getting to it's normal error
codes. This problem doesn't occur when we write a dummy value (0x0) to
the accelerometer's configuration register to leave it in stanby (but
then perform the same operations).