Fixed subtle but bad bug in app loader
Was previously passing up the end of the data section as the heap_base
to the app. However, the end of the data section is the beginning of the
BSS section, so the result was that malloc was overwriting itself.
Instead pass up the end of the BSS segment, which the correct start of
the uninitialized heap.