Fixes app loading error with relocated data
Fixes the bug pointed out by @ppanuto and @brghena in #51 where global
references to relocated data (e.g. strings in the text segment, or
dynamic globals) compiles with hardcoded incorrect pointer.

This adds an explicit section in the app's linker that pulls out such
pointers and adds that section to the load_info structure at the top of
the app. The kernel app loader then performs the same relocation as for
the GOT.