Scheduler: preserve PSR register after wait
The HW stacked PSR on exception entry includes a bit telling the HW
whether an aligner is present on the stack. We, therefore, need to
preserve this register after a call to wait (the scheduler pop's the HW
stack on wait) and restore it when we switch into a callback.

This register also includes information about the state of computations
(e.g. negative, overflow, carry flags), so best to preserve it anyway.