Add system calls for controlling app memory break
The new memory layout of an app is:

0xFF--> /--------\
        | KHEAP  |
        |   ||   |
        |   \/   |
        +--------+
         ////////
         ////////
        +--------+
        |   /\   |
        |   ||   |
        |  HEAP  |
        |________|
        | STACK  |
        |  BSS   |
        |  DATA  |
        |  GOT   |
0x0 --> \--------/

When the kernel allocates in the applications memory space, it moves the
kernel break down. The app needs to be able to move the memory break up
and down (e.g. in order to implement _sbrk).

We add the MEMOP system call with, currently, two operations:

0 is brk, which changes the application break to a given address
1 is sbrk, which moves the application break by a given diff

We will probably want to add operations that provide the app information
about current memory stats (how much heap the kernel is using, for
example)