Temporary fix for printf issue
Putstr is non-reentrant, which is a problem.

High level fix is to buffer putstrs, and in the callback for putstr,
invoke the next one.

Unfortunately, this requires wait_for to differentiate between different
calls to putstr, so we need unique values for each one we enqueue. Right
now using a hack that just increments a global counter and uses that counter as
the cb_type to wait for.