Sleeping seems to work from a state transition perspective. Haven't implemented DEEP_SLEEP, could do that if needed.
Changed the RF233 app to be awake for 10s, send, sleep for 1s. Did
this so it's likely two nodes will transmit when each other are awake
but time periods are long enough for even visual inspection of scope
traces. To increase speed modify the constants in main() of rf233.c.