Stack process's r4-r11 on kernel-side of syscall
For system calls this doesn't make much of a difference*, but it will
allow us to do preemption. Basically, preemption happens without the
process's knowledge, so we have to stack those registers in the kernel
anyway. This change allows us to treat both situations the same when
context switching back to the process.

* Except for technically it would be possible to overrun the process
stack by 32 bytes in the interrupt handler, while not running with the
MPU. We ought to deal with that somehow. A fault page won't work because
interrupt handlers are never subject to MPU rules.