Stack process r4-r11 in kernel
For system calls this doesn't make much of a difference*, but it will
allow us to do preemption more cleanly (basically making sure those
registers are stacked even if the context switch happened not because of
a system call).

* Except for technically it would be possible to overrun the process
stack by 32 bytes in the kernel, which is not protected. We ought to
deal with that...