Generic ISR handelr for IRQs
The handler is in assembler and jumps to the approriate ISR from a
global table. Using such a wrapper preps us for adding a prologue that
switches to the kernel after every IRQ.

If we had naked functions in Rust we may not need this (could just
implement it as part of the handler macro), which would save some
memory. Newer versions of Rust have naked function support, so should
explore then.