Timer driver: fix fired logic to account for freq.
timer.interval is in milliseconds, but elapsed is computed in clock
cycles. Account for this by computing the interval value in clock cycles
using the Frequency type. Also need to account for some jitter which may
lead to the alarm firing slightly before expected (seems to be by one
clock cycle, but just using 100 now to be safe, that's probably the
wrong value)

Also added some comments inline to make it clearer what we're doing