Create new non-timer based blink app
Now that timeslicing is up and running, there is no harm in having a
synchronous blink app and it is useful for porting new hardware (before
a driver for the hardware driver has been implemented).

Renames the timer-based, asynchronous, blink app to "blink" and creates
a new synchronous app (that just spins for a while) called "blink_sync".