Add no_mangle to ensure interrupt vectors aren't removed by rustc
The code appears to have been broken in rust-lang/rust#30830. This
commit uses no_mangle as a work-around, since that's special-cased by
the compiler and always kept.

A more proper solution would be to re-export the necessary symbols in
the final crate, which should also ensure that they're kept. This would
either introduce a platform dependency in the main crate, or require a
restructuring of the code so that the kernel binary is produced by the
platform code.