Reverse dependency between platform and main
In the new world, the `platform` crate depends is the main crate and it
depends on the `main` crate (which includes the scheduler). The platform
defines a reset handler for use by the chip's vector table. The
reset_handler calls the kernel `main` function (i.e. the scheduler)
when it's done initializing the platform.

This required some other cleanup that we wanted anyway, e.g., moving
systick into the architecture crate instead of a platform one, where it
belongs.

The changes are fairly wide reaching, but mostly simple.