Funge NRF51 port for Cargo to work with Makefile
Mostly little changes. The major difference is we have to compile
kernel.elf directly with Rust otherwise the symbols that used to be in
ctx_switch.S are not visible. This, in turn, affects how we incorporate
apps into the a final product -- specifically by using objcopy instead
of compiling.