Shift from 10001 to 10028. Timer worked once, now breaking consistently in blink_periodic.
The bug in this commit seems to be that Container memory is being
corrupted between calls and their callbacks. Specifically, if you
load apps/blink_periodic onto an nRF PCA10028, then you'll see this
behavior:

LED 1 blinks to show system boots.
LED 1 turns on to show main loop executes.
      (apps/blink_periodic/main.c:60)
LED 2 turns on to show that the call to subcribe to the timer event
      executes correctly.
      (src/drivers/timer.rs:65).
LED 3 turns on to show that the callback has been called from
      the hardware alarm into the timer driver.
      (src/drivers/timer.rs:157)
LED 4 turns on to show that the application has received the callback.
      (apps/blink_periodic/main.c:45)
LED 1 turns off to show main loop executes.
      (apps/blink_periodic/main.c:60)
LED 2 turns off to show that the call to subcribe to the timer event
      executes correctly.
      (src/drivers/timer.rs:65)
LED 3 DOES NOT turn off, because the if statement on
      src/drivers/timer.rs:64 fails (the application ID structure has
      been corrupted. If you comment out this if statement, then LED 3
      will turn off but LED 4 will not because the callback can't
      be correcrtly delivered to the application.