Add clock init function
This sets the SAM4L to the DFLL clock based on RC32K with the cache and
high speed flash enabled. It's basically a direct copy of the
sysclk_init() function from the ASF.

It also checks to see if the DFLL has already been configured, so it
works if the UART bootloader is present.