Remove layout symlinks
After discussing with @bradjc and @brghena, there's not a lot of value
in having the layout files attached to the chip, especially if they're just
marking memory ranges.

If you do build a platform with the same chip, copying memory ranges is
a minimal amount of duplicated content, and more to the point not something
that you really want to change out from underneath you if something in the
default layout provided by Tock changes.