Use package name for IPC  discovery mechanism
Allows a process to look up an service id by package name using the
allow system call.

An allow system call with target_id zero means a service lookup and will
return the service id if any of the processes package names
match the AppSlice passed into allow, or an error (-1) if none match.

An implication is that we change all IPC service number to be the
process id + 1 to accommodate a special meaning for 0 target id.