Adapt rot13 svc & client exmaple to IPC discovery
Uses service discovery instead of hard-coded IPC service ids in the
rot13 service/client example processes.

Currently relies on the service's package name's length being a multiple
of eight so that the actual binary (which appears directly after it)
remains eight-byte aligned. One solution could be to pad the package
name in the binary, but probably easiest just to put it at the end of
the binary instead. This fix will come in the next commit.