SAM4L ADC: Fix multiple interrupt issue (#204)
For some reason the ADC interrupt handler is firing twice for every adc
sample. I checked the status register, and the result is 0x51000001,
where the top 3 1s indicated various things are enabled, and the LSB is
one indicating that the conversion is finished. The status register on
the second interrupt is just 0x51000000.

What we really care about is not calling the callback twice, and this
ensures we only call the callback on the correct interrupt.

Also, `result.fired` needs to be cleared in the userland ADC library.