ISL29035: Add timer to wait for integration (#203)
Based on the ADC resolution, there is an integration time between when
a measurement is started and when the result is ready. This adds a timer
to wait for that amount of time.

I believe this isn't an issue with I2C at 100 kHz, as the latency of
writing bytes over I2C was long enough that the 410 us waiting period
was met. With faster I2C speeds, however, this is necessary. It will
also be necessary if a 12 bit or 16 bit conversion is used in the
future.

Fixes #165