travis: Only build mac on master
If there is a better way to do this I don't know what it is. Travis
makes it difficult to customize builds in this way. This should build
linux all the time (like before) but save the slower mac build for only
when a commit gets to master so we can keep an eye on the mac build,
without having to wait for it on every commit. If the mac build built
different code that would be one thing, but as it stands now it should
generate the exact same output, so we don't need to check it as often.