Up userland max app size to 16384
So, there's an issue here that demands more thought. Now that processes
can be any size, there's not a great way to check at compile time what the
limit on process size should be.

In theory, we could up this to be any size and only detect at load time that
things don't fit. In practice, that kinda sucks.

The process size commits up'd the heap from 1K to 2K by default, which caused
some of the bigger apps not to fit in 8K. Unclear what the best defaults for
all that stuff is either. This whole area demands more exploration.