kernel: round total app size up to power of two
This is likely a temporary fix that makes the grant region for the app
on a valid address so that the MPU region is properly aligned. There is
probably somewhere else that provides enough space while also being
correctly aligned that we should be using, but this makes Tock work
again.