Fixed off-by-1 error in packet length.
radio::HEADER_SIZE tells a caller what offset the payload should
be at. However, in 802.15.4, the packet length does not include the
length field itself. Additionally, 2 more bytes are needed for the
frame check sequence (FCS) aka CRC. The total header length is 10 bytes

So this means that if an application
wants to send 10 bytes of data, it should pass a frame where the data
is in bytes [10..19] and which is at least 22 bytes long, and the
length field will be 21 (9 (header not incl length field) + 10 (payload)
+ 2 (CRC/FCS).