Fixed RX/TX bugs. Two transcievers can now interleave transmissions.
Have not debugged what happens when a frame arrives as one is being
written out. If the write needs to abort, the worst case the RF233
will need a lock and perform byte-by-byte operations. Or it can discard
the received frame. Or, it can find a way to cancel the DMA?