Print the yield_pc in debug trace as well
Sometimes the PC is meaningful, and sometimes it's garbage. i.e. if
the user application issues a `pop {pc}` with trash on the stack,
then the PC printed in the debug handler will be the trash from the
stack, but the yeild_pc will point to the actual problem instruction.

There may be a way to figure this out automatically, but for the short
term printing both is an easy fix.