virtual_alarm: fix checking for fired timers
In certain cases where there are multiple virtual timers and events line
up just right, certain timers could be missed. To handle wrapping
concerns, timers are compared against a "previous" timestamp, however
this "previous" value can change while the virtual timers are being
checked. This can cause a valid timer's "when" time to be less than the
new previous value and the timer won't fire.

If I'm not mistaken, this issue happens when three things are true:

1. Two virtual timers are set to fire in the same 100 tic window.
2. One of the timers is from timer.rs.
3. The timer.rs timer is earlier in the array of multiplexed timers.

Because timer.rs will reschedule a timer if it fires too early (because
of the window) it calls `set_timer()` and `self.mux.prev` is updated to
a new time.

Checking all timers by the same "previous" value should fix this.