Fixes overflow bug in RNG driver if buf > num
If `remaining` was smaller than buffer size, we could end up in a
situation where remaining overflowed and was super big (close to
MAX_INT). Eventually, idx would reach the buffer size, so we were
operating on an empty buffer, but remaining was super big. The copy loop
wouldn't do anything (because the buffer seems empty), but remaining
would never go to zero.