Catch and reject null callback
Rust's NonZero "function" is really just a compiler hint that the
value enclosed by it is guarenteed not to be zero. Critically, NonZero
doesn't check this, it trusts that the caller validated this in advance
(which is why it's an unsafe fn). We were violating this assumption.

My first instinct was to just rip out the NonZero, because it wasn't
really buying us much - this is just a pointer that rust is holding
onto. However, removing it caused every App struct that holds a
Callback to grow by 8 bytes, even still if you change the type of
`fn_ptr` to usize instead of a function (pointer), so clearly Rust's
optimizer is doing something magical and it's worth keeping around.