Sketch out MapCell type
Starts to address Issue #178 where TakeCell actually moves data out of
the option. This is a performance problem if the data in a TakeCell is
large, but saves both cycles and memory if the data is small (like a
pointer or slice) because the Option can be optimized in many cases and
the code operates on registers as opposed to memory.

To resolve this, just introducing a new Cell type called MapCell which
keeps the data in the MapCell during a `map` operation. Otherwise the
interface is the same so should be a dropin replacement.