RF233 stack as Rust capsule (#261)
* First stab at an RF233 driver in Rust. Committing to make Sized issue
visible.

* Make the userland SPI driver be a client of a virtualized SPI.
This means virtualized SPIs need to be able to set the chip select
line, which results in a bunch of changes to the virtualized interface.

* Getting RF233 starting. 29 states!

* Include constants.

* RF233 compiling again. Next step: state machine.

* Compiling initialization sequence. Not tested yet.

* Fix starvation bug in virtualized SPI were pending operations handled
after new ones.

Fixed virtualized SPI idle state set bug where invoking operations in
a callback would fail because the operation would be later set to idle.

Changed RF233 to always read/write for registers to ensure chip select
line is cleared between each register operation (see GitHub issue #231).

* SPI capsule now working for buffer operations. Three bugs:
 - System call identifiers had changed but libtock hadn't been updated
 - Needed to configure buffers for capsule
 - Needed to set up callback for capsule.

* Can now initialize the RF233 and handle interrupts. PLL_ON seems to
not be triggering an interrupt, however, will compare with traces
from userland.

* Enable chip select in the virtualized SPI, remove debugging pin calls.

* PLL_LOCK is 1 << 0, not 0. Makes sense why PLL lock
interrupts not being delivered!

* Add RX_ON state.

* RF233 seemingly transmitting packets -- need to verify packets with a sniffer.

* Add radio HIL.

* Make setting the address and PAN ID push to hardware as well.

* Packet transmission working correctly.

* Add default address and PAN ID (0x1008 and 0xABCD). Fix endianness bug
in header generation.

* Working on receive path.

* Fixed off-by-1 error in packet length.

radio::HEADER_SIZE tells a caller what offset the payload should
be at. However, in 802.15.4, the packet length does not include the
length field itself. Additionally, 2 more bytes are needed for the
frame check sequence (FCS) aka CRC. The total header length is 10 bytes

So this means that if an application
wants to send 10 bytes of data, it should pass a frame where the data
is in bytes [10..19] and which is at least 22 bytes long, and the
length field will be 21 (9 (header not incl length field) + 10 (payload)
+ 2 (CRC/FCS).

* Fixed RX/TX bugs. Two transcievers can now interleave transmissions.

Have not debugged what happens when a frame arrives as one is being
written out. If the write needs to abort, the worst case the RF233
will need a lock and perform byte-by-byte operations. Or it can discard
the received frame. Or, it can find a way to cancel the DMA?

* First stab at interface for receiving/sending packets.

Note: implementation broken and unsafe due to unsafe{} code.

* Clean up warnings in code.

* Add length sanity checks.

* Fixed buffer handling, radio now receives/transmits with only one
copy (application to driver).

* Can now transmit packets from userland.

* Clean up warnings, add a few comments on buffer allocation.

* Reception working and delivering to userland, but an array offset
bug is hard faulting.

* Userland transmission and reception now working. There is a bug I
can't figure out. On a successful transmission, the TX command
system call should return 0 (radio.rs:208). However, it doesn't;
the userland library sees it return non-zero. This causes tickles
a bug in which an immediate call to receive causes a hard fault.
Specifically, I see this:

packet 1 received successfully
packet 1 echoed in transmission successfully
packet 2 received successfully
packet 2 transmission request
userland library tries to call receive again (because it doesn't
  wait for tx completion callback) and there's a hard fault.

I can't get gdb working on imix. So I'm punting on this for now,
but will come back to it. The current fix is that the userland
library always waits for a callback, even if the TX request fails.
This means passing bad parameters can cause your call to block
forever.

* Add some high-level comments.

* Radio capsule.

* rustfmt

* Some more comments.

* More accurate comments.

* Reconcole scheduler

* Reconcole callback too.

* Slightly better return values.

* Fix comment formatting.

* Synchronous acknowledgements now work. Need to incorporate into
transmit path so send done denote whether ack received. Note:
the header was being formatted incorrectly; FCF byte 0 should
be 0x88 not 0xAA; 0xAA gives a Reserved value for the Frame Version
so does not pass address recognition on the RF233.

* Fix commenty

* Addressing Pat's comments.

* Incorporating Pat's comments.

* rustfmt