MapCell (#271)
* Sketch out MapCell type

Starts to address Issue #178 where TakeCell actually moves data out of
the option. This is a performance problem if the data in a TakeCell is
large, but saves both cycles and memory if the data is small (like a
pointer or slice) because the Option can be optimized in many cases and
the code operates on registers as opposed to memory.

To resolve this, just introducing a new Cell type called MapCell which
keeps the data in the MapCell during a `map` operation. Otherwise the
interface is the same so should be a dropin replacement.

* Benchmark for TakeCell vs MapCell

Utility test in the `hail` board that measures memory overhead of
TakeCell and MapCell, as well as cycles for a no-op map operation on
data of increasing sizes:

| Type         |Cycles | Bytes|
|--------------|-------|------|
|`&u32`        | 3     | 4    |
|`[[u8;256];1]`| 1655  | 257  |
|`[[u8;256];2]`| 2795  | 513  |
|`[[u8;256];3]`| 4365  | 769  |
|`[[u8;256];4]`| 4838  | 1025 |
|`[[u8;256];5]`| 6780  | 1281 |
|`[[u8;256];6]`| 7706  | 1537 |
|`[[u8;256];7]`| 10326 | 1793 |

| Type         | Cycles | Bytes|
|--------------|--------|------|
|`&u32`        | 18     | 8    |
|`[[u8;256];1]`| 20     | 257  |
|`[[u8;256];2]`| 22     | 513  |
|`[[u8;256];3]`| 20     | 769  |
|`[[u8;256];4]`| 22     | 1025 |
|`[[u8;256];5]`| 22     | 1281 |
|`[[u8;256];6]`| 22     | 1537 |
|`[[u8;256];7]`| 20     | 1793 |

* Convert TakeCell to MapCell/Cell if appropriate

Changes TakeCell to only admit mutable references. Other types should
simply use Cell or MapCell.

* docs

* rustfmt

* rustfmt