Two small updates to SPI over USART
 * Make the ChipSelect an Option<GPIO>>, allowing drivers to use the RTS line if desired
 * Make the SPI transaction end on DMA RX complete, when applicable

The second bullet goes part-way to fixing issue #274. In the case of a
SPI transaction that does a read, the CS line will now be deasserted
properly. In the case of a many-byte write with no read, the problem is
not solved (but has not been made worse either).

This has been tested and works on the signpost platform which uses FRAM
over SPI over USART including both read/writes and write-only transfers.