userland: per-app region size handling
The way this was originally done was always a bit of a hack. For some
reason, the first time I did this I couldn't get passing variables to
the linker to work correctly. This is now a much cleaner implmentation
and will size regions for each app instead of tieing it to libtock.