Add an imix benchmark configuration
Adds a new board configuration for imix with no drivers at all, and just
the minimal setup to get the kernel running and pins configured for
various benchmarks.

Adds a first benchmark that measures the latency in a capsule between an
external GPIO trigger (pressing the user button) and an external GPIO
output (toggling the LED).