dynamic userland SRAM sizes
The changes the build to dynamically size the SRAM and FLASH regions
in the userland layout to match the underlying board.

A side-effect of this change is that find build artifacts now necessarily
live in `build/TOCK_BOARD/`. Most of the built artifacts remain shared
under `build/TOCK_ARCH/`, it's only the link and post-link artifacts that
live in the board-specific dir.