don't make assumptions about Tock ram footprint
As brought up by @niklasad1 on IRC, the conservative Tock footprint
estimate is troublesome for platforms with tight memory constraints,
like the nrf.

It's not really a problem per se for an app to build when there's no
enough memory for it to fit, really just an inconvenience as learning
that now requires flashing the app and detecting the size failure at
runtime.