kernel+sam4l: provide CRC hil + implement CRCCU
N.B.: This commit performs a debug routine for the CRCCU in the imix
initialization code.

This is a Technology Preview, meaning it doesn't work.

The crc_test_begin() function in imix/src/crc_dummy.rs attempts to
perform a CRC computation once the imix has started up.

It was expected that the CRCCU would issue an interrupt upon completion
of the calculation (as requested by writing to the DMAIER register), but
none comes.  (If the CRCCU interrupt fires, the dummy code will loop
while flashing the LED.)

If instead one does not wait for the interrupt but simply busy-waits,
the unit *does* leave a value in SR and a read of DMASR indicates the
DMA transfer has completed.  (There is commented-out code in crccu.rs to
experiment with this.)

Also, if one configures "compare" mode, the unit *does* issue an
interrupt indicating that the resulting value differs from the
configured reference value.