Fix SAM4L AST bug when delay is very short
We've been seeing issues using the AST when setting an alarm to very
short (a few tics) future clock values.

The datasheet has the following ominous language (Section 19.5.3.2):

> Because of synchronization, the transfer of the alarm value will not
> happen immediately. When changing/setting the alarm value, the user must
> make sure that the counter will not count the selected alarm value before
> the value is transferred to the register. In that case, the first alarm
> interrupt after the change will not be triggered.

In practice, we've observed that when the alarm is set for a counter value
less than or equal to four tics ahead of the current counter value, the
alarm interrupt doesn't fire.

This commit simply rounds up to at least eight tics, which seems safe
enough and in practice has seemed to work. It also optimizes out the
clock enable/disable during `set_timer` which were doing more harm than
good.