hail: update bootloader
Update the binary in the hail folder with the correct bootloader and
update the Makefile to let you program it with tockloader.

Ideally this will go away. There is no real reason that the bootloader
binary is in tock, but we keep it for now as things are changing so that
people can easily update their hails.