imix v2.0+: use tockloader/bootloader, not openocd
Due to a hardware fix to the serial port, imix v2.0 and onward will be
compatible with tockloader and the tock-bootloader. This eliminates an
external dependency on openocd.

The tock-bootloader was configured for imix (and a pull request
submitted to that repo). The Makefile, README, and chip_layout were
basically copied from hail.

Tested successfully with a modified imix v1.3 and tockloader. All
tockloader features worked, and apps ran correctly.