Add a common type representing powers of two
This is a simple wrapper type representing numbers which are an exact
power of two. Under the hood, they are simply represented as a `u32`
of the base-two exponent. This will allow us to pass around values
guaranteed to be exact powers of two without having to reverify them
everywhere.