kernel: Update to TBF Header v2
Tock Binary Format header version 2 adds a `flags` field that can be
used to mark properties of a flashed application.

The two supported flags are:

1. Enable/disable. Tells the kernel to run or not run an application at
boot.
2. Sticky. Tells tockloader to not erase the app without --force.

It also allows apps to tell the kernel to not do PIC fixups on its
behalf.

Apps can also now specify flash regions that the app intends to write.

Fixes #379